refactory: review 도메인 이동
