--- backend/src/main/java/toonpick/app/dto/RecommendationListDTO.java
@@ -1,14 +1,0 @@
-package toonpick.app.dto;
-
-import lombok.Data;
-
-import java.util.List;
-
-
-@Data
-public class RecommendationListDTO {
-    private Long id;
-    private String theme;
-    private String description;
-    private List<Long> webtoonIds;
-}
