+++ backend/src/main/java/toonpick/app/review/entity/ReviewLike.java
@@ -1,4 +1,4 @@
-package toonpick.app.entity;
+package toonpick.app.review.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
