+++ backend/src/main/java/toonpick/app/review/entity/WebtoonReview.java
@@ -1,10 +1,11 @@
-package toonpick.app.entity;
+package toonpick.app.review.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import toonpick.app.entity.BaseTimeEntity;
 import toonpick.app.member.entity.Member;
 import toonpick.app.webtoon.entity.Webtoon;
