+++ backend/src/test/java/toonpick/app/unit/service/WebtoonReviewServiceTest.java
@@ -5,18 +5,18 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-import toonpick.app.dto.WebtoonReviewCreateDTO;
-import toonpick.app.dto.WebtoonReviewDTO;
+import toonpick.app.review.dto.WebtoonReviewCreateDTO;
+import toonpick.app.review.dto.WebtoonReviewDTO;
 import toonpick.app.member.entity.Member;
 import toonpick.app.webtoon.entity.Webtoon;
-import toonpick.app.entity.WebtoonReview;
+import toonpick.app.review.entity.WebtoonReview;
 import toonpick.app.entity.enums.Platform;
 import toonpick.app.entity.enums.SerializationStatus;
 import toonpick.app.exception.ResourceNotFoundException;
 import toonpick.app.member.repository.MemberRepository;
 import toonpick.app.webtoon.repository.WebtoonRepository;
-import toonpick.app.repository.WebtoonReviewRepository;
-import toonpick.app.service.WebtoonReviewService;
+import toonpick.app.review.repository.WebtoonReviewRepository;
+import toonpick.app.review.service.WebtoonReviewService;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
